8月8号（何金钊 药品调拨出库）

+++ src/main/java/cn/gson/hui_ren_boot/model/pojos/pharmacy/PlanDetails.java
@@ -10,7 +10,7 @@ public class PlanDetails {
 
     private long planId;
     private String planSerial;
-    private Long purchasePrice;
+    private double purchasePrice;
     private String drugId;
     private Long supplierId;
     private Long planNum;
@@ -44,11 +44,11 @@ public class PlanDetails {
 
     @Basic
     @Column(name = "PURCHASE_PRICE")
-    public Long getPurchasePrice() {
+    public double getPurchasePrice() {
         return purchasePrice;
     }
 
-    public void setPurchasePrice(Long purchasePrice) {
+    public void setPurchasePrice(double purchasePrice) {
         this.purchasePrice = purchasePrice;
     }
